refactor: merge `HostedServer` into `JazzServer`
